+++ pypgx/__init__.py
@@ -53,6 +53,7 @@ from .api.utils import (
     predict_alleles,
     predict_cnv,
     prepare_depth_of_coverage,
+    print_data,
     print_metadata,
     slice_bam,
     test_cnv_caller,